requirements.txt
PyQt5==5.15.9
pyqtgraph
numpy
scipy
pyserial
matplotlib
loguru
pyyaml
